requirements.txt
requests
websocket-client
pyaudio
sounddevice
soundfile
beautifulsoup4
deepgram-sdk
groq
python-dotenv
lxml